refactor php-script.h
