+++ server/php-init-scripts.h
@@ -1,0 +1,30 @@
+// Compiler for PHP (aka KPHP)
+// Copyright (c) 2020 LLC «V Kontakte»
+// Distributed under the GPL v3 License, see LICENSE.notice.txt
+
+#pragma once
+
+#include <cstddef>
+
+struct script_t {
+  void (*run)();   // this is entrypoint to generated code
+  void (*clear)();
+};
+
+/// It binds generated code and runtime. Definition is generated by compiler. This function is called at global initialization @see init_all().
+void init_php_scripts() noexcept;
+/// It initializes const variables. Definition is generated by compiler. This function is called at global initialization @see init_all().
+void global_init_php_scripts() noexcept;
+
+script_t *get_script();
+/// It's called from init_php_scripts() and set the entrypoint to generated code
+void set_script(void (*run)(), void (*clear)());
+
+struct script_result {
+  const char *headers;
+  int headers_len;
+  const char *body;
+  int body_len;
+  int exit_code;
+};
+
