+++ server/php-runner.h
@@ -10,9 +10,9 @@
 #include "common/sanitizer.h"
 
 #include "server/php-engine-vars.h"
-#include "server/php-query-data.h"
+#include "server/php-init-scripts.h"
 #include "server/php-queries-types.h"
-#include "server/php-script.h"
+#include "server/php-query-data.h"
 #include "server/ucontext-portable.h"
 
 enum class run_state_t {
